Put shared function into one location for 'markTemplateAsDirty'
