+++ src/resolvers/question.ts
@@ -2,8 +2,9 @@ import { Resolvers } from "../types";
 import { MyContext } from "../context";
 import { QuestionOption } from "../models/QuestionOption";
 import { Question } from "../models/Question";
-import { getQuestionOptionsToRemove, markTemplateAsDirty } from "../services/questionService";
+import { getQuestionOptionsToRemove } from "../services/questionService";
 import { AuthenticationError, ForbiddenError, InternalServerError, NotFoundError } from "../utils/graphQLErrors";
+import { markTemplateAsDirty } from "../utils/templateUtils";
 import { QuestionCondition } from "../models/QuestionCondition";
 import { formatLogMessage } from "../logger";
 import { isAdmin, isAuthorized } from "../services/authService";
