+++ src/resolvers/section.ts
@@ -4,8 +4,9 @@ import { Section } from "../models/Section";
 import { VersionedSection } from "../models/VersionedSection";
 import { Tag } from "../models/Tag";
 import { Template } from "../models/Template";
-import { cloneSection, hasPermissionOnSection, markTemplateAsDirty } from "../services/sectionService";
+import { cloneSection, hasPermissionOnSection } from "../services/sectionService";
 import { ForbiddenError, NotFoundError, AuthenticationError, InternalServerError } from "../utils/graphQLErrors";
+import { markTemplateAsDirty } from "../utils/templateUtils";
 import { Question } from "../models/Question";
 import { isAdmin, isAuthorized, isSuperAdmin } from "../services/authService";
 import { formatLogMessage } from "../logger";
