+++ src/services/__tests__/templateService.spec.ts
@@ -369,7 +369,7 @@ describe('template versioning', () => {
     expect(updated.modifiedById).toEqual(tmplt.modifiedById);
     expect(updated.modified).toEqual(tmplt.modified);
     expect(updated.latestPublishVersion).toEqual(newVersion.version);
-    expect(updated.isDirty).toEqual(false);
+    expect(updated.isDirty).toEqual(true);
   });
 
   it('versions the Template when there are prior versions', async () => {
@@ -434,6 +434,6 @@ describe('template versioning', () => {
     expect(updated.modifiedById).toEqual(tmplt.modifiedById);
     expect(updated.modified).toEqual(tmplt.modified);
     expect(updated.latestPublishVersion).toEqual(newVersion.version);
-    expect(updated.isDirty).toEqual(false);
+    expect(updated.isDirty).toEqual(true);
   });
 });
