+++ src/services/questionService.ts
@@ -182,17 +182,3 @@ export const getQuestionOptionsToRemove = async (questionOptions: QuestionOption
 
   return Array.isArray(optionsToRemove) ? optionsToRemove : [];
 }
-
-// Template needs to be updated to isDirty=true if changes were made to its sections or questions
-export const markTemplateAsDirty = async (reference: string, context: MyContext, templateId: number): Promise<void> => {
-  try {
-    const template = await Template.findById(reference, context, templateId);
-    if (template) {
-      template.isDirty = true;
-      await template.update(context);
-    }
-  } catch (err) {
-    formatLogMessage(context).error(err, `Failure in ${reference} - markTemplateAsDirty`);
-    throw err;
-  }
-};
