+++ src/services/sectionService.ts
@@ -116,17 +116,3 @@ export const hasPermissionOnSection = async (context: MyContext, templateId: num
   // Offload permission checks to the Template
   return await hasPermissionOnTemplate(context, template);
 }
-
-// Template needs to be updated to isDirty=true if changes were made to its sections or questions
-export const markTemplateAsDirty = async (reference: string, context: MyContext, templateId: number): Promise<void> => {
-  try {
-    const template = await Template.findById(reference, context, templateId);
-    if (template) {
-      template.isDirty = true;
-      await template.update(context);
-    }
-  } catch (err) {
-    formatLogMessage(context).error(err, `Failure in ${reference} - markTemplateAsDirty`);
-    throw err;
-  }
-};
